Update README.md
Changed the description of Far Horizons from:

FAR HORIZONS is a strategic role-playing game of galactic exploration, trade, diplomacy, and conquest.

To:

FAR HORIZONS is a strategic play-by-email game of galactic exploration, trade, diplomacy, and conquest.

"Role-playing" is not the best description of this game. It's more a strategic space 4x game than anything, but the fact that it's play-by-email is the best way to describe it.